- Add additional apiclient test and refactor reusable code into a mock utils class
